requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
anthropic>=0.40.0
httpx>=0.27.0
jinja2==3.1.3
aiofiles==23.2.1
python-multipart==0.0.6
